refactor: 상수 constants에 추가 및 import하도록 수정
