+++ frontend/src/constants/index.ts
@@ -3,3 +3,6 @@ export const BACKEND_URL = process.env.NEXT_PUBLIC_SOCKET_URI;
 export const HEADERS = { 'Content-Type': 'application/json' };
 export const SUCCESS = 'SUCCESS';
 export const FAILED = 'FAILED';
+export const ID_EMPTY_MSG = '아이디를 입력해 주세요';
+export const PASSWORD_EMPTY_MSG = '비밀번호를 입력해 주세요';
+export const NICKNAME_EMPTY_MSG = '닉네임을 입력해 주세요';
