+++ frontend/src/pages/findPwd/index.tsx
@@ -4,10 +4,11 @@ import styled from '@emotion/styled';
 import { NextPage } from 'next';
 import { useSelector } from 'react-redux';
 
-import { requestChangePassword, ID_EMPTY_MSG, PASSWORD_EMPTY_MSG, NICKNAME_EMPTY_MSG } from '~/api/account';
+import { requestChangePassword } from '~/api/account';
 import InputBox from '~/atoms/InputBox';
 import MenuInfoBox from '~/atoms/MenuInfoBox';
 import PageBox from '~/atoms/PageBox';
+import { ID_EMPTY_MSG, PASSWORD_EMPTY_MSG, NICKNAME_EMPTY_MSG } from '~/constants/index';
 import ResponsiveButton from '~/molecules/ResponsiveButton';
 import theme from '~/styles/theme';
