+++ frontend/src/pages/mypage/index.tsx
@@ -162,6 +162,7 @@ const MyPage: NextPage = () => {
   const changeBooduckColor = (newColor: string) => {
     setColor(() => {
       changeColor(id, newColor);
+    
       return newColor;
     });
   };
